move MockHttpServer to TestUtils; for reusage purposes
