+++ src/test/scala/io/hydrosphere/mist/master/TestUtils.scala
@@ -1,9 +1,11 @@
 package io.hydrosphere.mist.master
 
+import akka.http.scaladsl.model.{HttpRequest, HttpResponse}
+import akka.stream.scaladsl.Flow
 import com.typesafe.config.ConfigFactory
 
 import scala.concurrent.duration.Duration
-import scala.concurrent.{Await, Future}
+import scala.concurrent.{Await, Future, Promise}
 
 object TestUtils {
 
@@ -39,6 +41,50 @@ object TestUtils {
 
   implicit class AwaitSyntax[A](f: => Future[A]) {
     def await: A = Await.result(f, Duration.Inf)
+  }
+
+
+  object MockHttpServer {
+
+    import akka.actor.ActorSystem
+    import akka.http.scaladsl.Http
+    import akka.stream.ActorMaterializer
+    import akka.util.Timeout
+
+    import scala.concurrent.duration._
+
+    def onServer[A](
+      routes: Flow[HttpRequest, HttpResponse, Unit],
+      f: (Http.ServerBinding) => A): Future[A] = {
+
+      implicit val system = ActorSystem("mock-http-cli")
+      implicit val materializer = ActorMaterializer()
+
+      implicit val executionContext = system.dispatcher
+      implicit val timeout = Timeout(1.seconds)
+
+      val binding = Http().bindAndHandle(routes, "localhost", 0)
+
+      val close = Promise[Http.ServerBinding]
+      close.future
+        .flatMap(binding => binding.unbind())
+        .onComplete(_ => {
+          system.shutdown()
+          system.awaitTermination()
+        })
+
+      val result = binding.flatMap(binding => {
+        try {
+          Future.successful(f(binding))
+        } catch {
+          case e: Throwable =>
+            Future.failed(e)
+        } finally {
+          close.success(binding)
+        }
+      })
+      result
+    }
   }
 
 }
